Added song component
