+++ src/lib/components/Song.svelte
@@ -1,0 +1,27 @@
+<script lang="ts">
+	import type { File } from '$lib/types';
+
+	export let href: string;
+	export let file: File;
+
+	function missingImg(event: Event) {
+		const target = event.target as HTMLImageElement;
+		target.src = '/svelte.png';
+	}
+</script>
+
+<a {href} data-sveltekit-noscroll={file.type === 'folder' ? 'off' : ''}>
+	<div
+		class="flex items-center gap-x-4 rounded-md hover:shadow-xl transition-shadow pr-4 h-20 overflow-hidden"
+	>
+		<img
+			src={file.cover_url || '/svelte.png'}
+			alt="Cover"
+			on:error={missingImg}
+			class="aspect-square h-full object-cover"
+		/>
+		<p class="text-contrast">
+			{file.display_name}
+		</p>
+	</div>
+</a>
